Add shared team context provider
